# Update TrinoFileSystemCache to represent latest hadoop implementation